v2.4 ALPHA A
Patch to fix problem of some sample files sent to S950 can be written to floppy-disk emulator but can't be read back. Problem is that the S950 requires an even number of sample words for reading from disk. So .WAV files with an odd number of samples always failed to be read back from a floppy-drive emulator.